+++ src/Fonts/gfxfont.hpp
@@ -166,5 +166,70 @@ struct GFXfont : public lgfx::IFont { // Data stored for FONT AS A WHOLE:
 #include "GFXFF/FreeSerifBoldItalic18pt7b.h" // FF47 or FSBI18
 #include "GFXFF/FreeSerifBoldItalic24pt7b.h" // FF48 or FSBI24
 
+#include "Custom/Orbitron_Light_24.h"
+#include "Custom/Orbitron_Light_32.h"
+#include "Custom/Roboto_Thin_24.h"
+#include "Custom/Satisfy_24.h"
+#include "Custom/Yellowtail_32.h"
+
+namespace fonts {
+
+  static PROGMEM const GFXfont& TomThumb                  = TomThumb                 ;
+  static PROGMEM const GFXfont& FreeMono9pt7b             = FreeMono9pt7b            ;
+  static PROGMEM const GFXfont& FreeMono12pt7b            = FreeMono12pt7b           ;
+  static PROGMEM const GFXfont& FreeMono18pt7b            = FreeMono18pt7b           ;
+  static PROGMEM const GFXfont& FreeMono24pt7b            = FreeMono24pt7b           ;
+  static PROGMEM const GFXfont& FreeMonoBold9pt7b         = FreeMonoBold9pt7b        ;
+  static PROGMEM const GFXfont& FreeMonoBold12pt7b        = FreeMonoBold12pt7b       ;
+  static PROGMEM const GFXfont& FreeMonoBold18pt7b        = FreeMonoBold18pt7b       ;
+  static PROGMEM const GFXfont& FreeMonoBold24pt7b        = FreeMonoBold24pt7b       ;
+  static PROGMEM const GFXfont& FreeMonoOblique9pt7b      = FreeMonoOblique9pt7b     ;
+  static PROGMEM const GFXfont& FreeMonoOblique12pt7b     = FreeMonoOblique12pt7b    ;
+  static PROGMEM const GFXfont& FreeMonoOblique18pt7b     = FreeMonoOblique18pt7b    ;
+  static PROGMEM const GFXfont& FreeMonoOblique24pt7b     = FreeMonoOblique24pt7b    ;
+  static PROGMEM const GFXfont& FreeMonoBoldOblique9pt7b  = FreeMonoBoldOblique9pt7b ;
+  static PROGMEM const GFXfont& FreeMonoBoldOblique12pt7b = FreeMonoBoldOblique12pt7b;
+  static PROGMEM const GFXfont& FreeMonoBoldOblique18pt7b = FreeMonoBoldOblique18pt7b;
+  static PROGMEM const GFXfont& FreeMonoBoldOblique24pt7b = FreeMonoBoldOblique24pt7b;
+  static PROGMEM const GFXfont& FreeSans9pt7b             = FreeSans9pt7b            ;
+  static PROGMEM const GFXfont& FreeSans12pt7b            = FreeSans12pt7b           ;
+  static PROGMEM const GFXfont& FreeSans18pt7b            = FreeSans18pt7b           ;
+  static PROGMEM const GFXfont& FreeSans24pt7b            = FreeSans24pt7b           ;
+  static PROGMEM const GFXfont& FreeSansBold9pt7b         = FreeSansBold9pt7b        ;
+  static PROGMEM const GFXfont& FreeSansBold12pt7b        = FreeSansBold12pt7b       ;
+  static PROGMEM const GFXfont& FreeSansBold18pt7b        = FreeSansBold18pt7b       ;
+  static PROGMEM const GFXfont& FreeSansBold24pt7b        = FreeSansBold24pt7b       ;
+  static PROGMEM const GFXfont& FreeSansOblique9pt7b      = FreeSansOblique9pt7b     ;
+  static PROGMEM const GFXfont& FreeSansOblique12pt7b     = FreeSansOblique12pt7b    ;
+  static PROGMEM const GFXfont& FreeSansOblique18pt7b     = FreeSansOblique18pt7b    ;
+  static PROGMEM const GFXfont& FreeSansOblique24pt7b     = FreeSansOblique24pt7b    ;
+  static PROGMEM const GFXfont& FreeSansBoldOblique9pt7b  = FreeSansBoldOblique9pt7b ;
+  static PROGMEM const GFXfont& FreeSansBoldOblique12pt7b = FreeSansBoldOblique12pt7b;
+  static PROGMEM const GFXfont& FreeSansBoldOblique18pt7b = FreeSansBoldOblique18pt7b;
+  static PROGMEM const GFXfont& FreeSansBoldOblique24pt7b = FreeSansBoldOblique24pt7b;
+  static PROGMEM const GFXfont& FreeSerif9pt7b            = FreeSerif9pt7b           ;
+  static PROGMEM const GFXfont& FreeSerif12pt7b           = FreeSerif12pt7b          ;
+  static PROGMEM const GFXfont& FreeSerif18pt7b           = FreeSerif18pt7b          ;
+  static PROGMEM const GFXfont& FreeSerif24pt7b           = FreeSerif24pt7b          ;
+  static PROGMEM const GFXfont& FreeSerifItalic9pt7b      = FreeSerifItalic9pt7b     ;
+  static PROGMEM const GFXfont& FreeSerifItalic12pt7b     = FreeSerifItalic12pt7b    ;
+  static PROGMEM const GFXfont& FreeSerifItalic18pt7b     = FreeSerifItalic18pt7b    ;
+  static PROGMEM const GFXfont& FreeSerifItalic24pt7b     = FreeSerifItalic24pt7b    ;
+  static PROGMEM const GFXfont& FreeSerifBold9pt7b        = FreeSerifBold9pt7b       ;
+  static PROGMEM const GFXfont& FreeSerifBold12pt7b       = FreeSerifBold12pt7b      ;
+  static PROGMEM const GFXfont& FreeSerifBold18pt7b       = FreeSerifBold18pt7b      ;
+  static PROGMEM const GFXfont& FreeSerifBold24pt7b       = FreeSerifBold24pt7b      ;
+  static PROGMEM const GFXfont& FreeSerifBoldItalic9pt7b  = FreeSerifBoldItalic9pt7b ;
+  static PROGMEM const GFXfont& FreeSerifBoldItalic12pt7b = FreeSerifBoldItalic12pt7b;
+  static PROGMEM const GFXfont& FreeSerifBoldItalic18pt7b = FreeSerifBoldItalic18pt7b;
+  static PROGMEM const GFXfont& FreeSerifBoldItalic24pt7b = FreeSerifBoldItalic24pt7b;
+
+  static PROGMEM const GFXfont& Orbitron_Light_24         = Orbitron_Light_24        ;
+  static PROGMEM const GFXfont& Orbitron_Light_32         = Orbitron_Light_32        ;
+  static PROGMEM const GFXfont& Roboto_Thin_24            = Roboto_Thin_24           ;
+  static PROGMEM const GFXfont& Satisfy_24                = Satisfy_24               ;
+  static PROGMEM const GFXfont& Yellowtail_32             = Yellowtail_32            ;
+
+}
 
 #endif
