+++ src/lgfx/lgfx_base.hpp
@@ -35,31 +35,11 @@ Contributors:
 
 namespace lgfx
 {
-  enum attribute_t
-  { cp437_switch = 1
-  , utf8_switch  = 2
-  };
-
-  static PROGMEM const IFont* fontdata [] = {
-    &fonts::font0,  // GLCD font (Font 0)
-    &fonts::font0,  // GLCD font (or GFX font)
-    &fonts::font2,
-    &fonts::font0,  // Font 3 current unused
-    &fonts::font4,
-    &fonts::font0,  // Font 5 current unused
-    &fonts::font6,
-    &fonts::font7,
-    &fonts::font8,
-  };
-
-//----------------------------------------------------------------------------
-
-//----------------------------------------------------------------------------
   class LGFXBase
   {
   public:
     LGFXBase() {}
-    virtual ~LGFXBase() { unloadFont(); }
+    virtual ~LGFXBase() {}
 
 // color param format:
 // rgb888 : uint32_t
@@ -123,7 +103,6 @@ namespace lgfx
 
     __attribute__ ((always_inline)) inline int32_t width        (void) const { return _width; }
     __attribute__ ((always_inline)) inline int32_t height       (void) const { return _height; }
-    __attribute__ ((always_inline)) inline const IFont* getFont (void) const { return _font; }
     __attribute__ ((always_inline)) inline color_depth_t getColorDepth(void) const { return _write_conv.depth; }
     __attribute__ ((always_inline)) inline color_conv_t* getColorConverter(void) { return &_write_conv; }
     __attribute__ ((always_inline)) inline bool hasPalette    (void) const { return _palette_count; }
@@ -1223,391 +1202,6 @@ namespace lgfx
       endWrite();
     }
 
-//----------------------------------------------------------------------------
-// print & text support
-//----------------------------------------------------------------------------
-// Arduino Print.h compatible
-    size_t print(const char str[])      { return write(str); }
-    size_t print(char c)                { return write(c); }
-    size_t print(int  n, int base = 10) { return print((long)n, base); }
-    size_t print(long n, int base = 10)
-    {
-      if (base == 0) { return write(n); }
-      if (base == 10) {
-        if (n < 0) {
-          size_t t = print('-');
-          return printNumber(-n, 10) + t;
-        }
-        return printNumber(n, 10);
-      }
-      return printNumber(n, base);
-    }
-
-    size_t print(unsigned char n, int base = 10) { return print((unsigned long)n, base); }
-    size_t print(unsigned int  n, int base = 10) { return print((unsigned long)n, base); }
-    size_t print(unsigned long n, int base = 10) { return (base) ? printNumber(n, base) : write(n); }
-    size_t print(double        n, int digits= 2) { return printFloat(n, digits); }
-
-    size_t println(void) { return print("\r\n"); }
-    size_t println(const char c[])                 { size_t t = print(c); return println() + t; }
-    size_t println(char c        )                 { size_t t = print(c); return println() + t; }
-    size_t println(int  n, int base = 10)          { size_t t = print(n,base); return println() + t; }
-    size_t println(long n, int base = 10)          { size_t t = print(n,base); return println() + t; }
-    size_t println(unsigned char n, int base = 10) { size_t t = print(n,base); return println() + t; }
-    size_t println(unsigned int  n, int base = 10) { size_t t = print(n,base); return println() + t; }
-    size_t println(unsigned long n, int base = 10) { size_t t = print(n,base); return println() + t; }
-    size_t println(double        n, int digits= 2) { size_t t = print(n, digits); return println() + t; }
-
-  #if defined (ARDUINO)
-    size_t print(const String &s) { return write(s.c_str(), s.length()); }
-    size_t print(const __FlashStringHelper *s)   { return print(reinterpret_cast<const char *>(s)); }
-    size_t println(const String &s)              { size_t t = print(s); return println() + t; }
-    size_t println(const __FlashStringHelper *s) { size_t t = print(s); return println() + t; }
-  #endif
-
-    size_t printf(const char * format, ...)  __attribute__ ((format (printf, 2, 3)))
-    {
-      char loc_buf[64];
-      char * temp = loc_buf;
-      va_list arg;
-      va_list copy;
-      va_start(arg, format);
-      va_copy(copy, arg);
-      size_t len = vsnprintf(temp, sizeof(loc_buf), format, copy);
-      va_end(copy);
-
-      if (len >= sizeof(loc_buf)){
-        temp = (char*) malloc(len+1);
-        if (temp == nullptr) {
-          va_end(arg);
-          return 0;
-        }
-        len = vsnprintf(temp, len+1, format, arg);
-      }
-      va_end(arg);
-      len = write((uint8_t*)temp, len);
-      if (temp != loc_buf){
-        free(temp);
-      }
-      return len;
-    }
-
-    size_t write(const char* str)                 { return (!str) ? 0 : write((const uint8_t*)str, strlen(str)); }
-    size_t write(const char *buf, size_t size)    { return write((const uint8_t *) buf, size); }
-    size_t write(const uint8_t *buf, size_t size) { size_t n = 0; startWrite(); while (size--) { n += write(*buf++); } endWrite(); return n; }
-    size_t write(uint8_t utf8)
-    {
-      if (utf8 == '\r') return 1;
-      if (utf8 == '\n') {
-        _filled_x = 0;
-        _cursor_x = 0;
-        _cursor_y += _font_size.y_advance * _textsize_y;
-      } else {
-        uint16_t uniCode = utf8;
-        if (_utf8) {
-          uniCode = decodeUTF8(utf8);
-          if (uniCode < 32) return 1;
-        }
-        //if (!(fpUpdateFontSize)(this, uniCode)) return 1;
-        if (!_font->updateFontMetric(&_font_size, uniCode)) return 1;
-
-        if (0 == _font_size.x_size) return 1;
-
-        int16_t w  = _font_size.x_size    * _textsize_x;
-        int16_t xo = _font_size.x_offset  * _textsize_x;
-        int16_t h  = _font_size.y_size    * _textsize_y;
-        int16_t yo = _font_size.y_offset  * _textsize_y;
-        if (_textscroll || _textwrap_x) {
-          int32_t left = _textscroll ? _sx : 0;
-          if (_cursor_x < left - xo) _cursor_x = left - xo;
-          else {
-            int32_t right = _textscroll ? _sx + _sw : _width;
-            if (_cursor_x + xo + w > right) {
-              _filled_x = 0;
-              _cursor_x = left - xo;
-              _cursor_y += _font_size.y_advance * _textsize_y;
-            }
-          }
-        }
-        if (_textscroll) {
-          if (_cursor_y < _sy - yo) _cursor_y = _sy - yo;
-          else {
-            int yshift = (_sy + _sh) - (_cursor_y + yo + h);
-            if (yshift < 0) {
-              scroll(0, yshift);
-              _cursor_y += yshift;
-            }
-          }
-        } else if (_textwrap_y) {
-          if (_cursor_y + yo + h > _height) {
-            _filled_x = 0;
-            _cursor_x = - xo;
-            _cursor_y = - yo;
-          } else
-          if (_cursor_y < - yo) _cursor_y = - yo;
-        }
-//      _cursor_x += (fpDrawChar)(this, _cursor_x, _cursor_y, uniCode, _text_fore_rgb888, _text_back_rgb888, _textsize_x, _textsize_y, fontdata[_textfont]);
-        _cursor_x += (fpDrawChar)(this, _cursor_x, _cursor_y, uniCode, _text_fore_rgb888, _text_back_rgb888, _textsize_x, _textsize_y, _font);
-      }
-
-      return 1;
-    }
-
-    int32_t textWidth(const char *string)
-    {
-      if (!string) return 0;
-
-      int32_t left = 0;
-      int32_t right = 0;
-      do {
-        uint16_t uniCode = *string;
-        if (_utf8) {
-          do {
-            uniCode = decodeUTF8(*string);
-          } while (uniCode < 32 && *(++string));
-          if (uniCode < 32) break;
-        }
-        //if (!(fpUpdateFontSize)(this, uniCode)) continue;
-        if (!_font->updateFontMetric(&_font_size, uniCode)) continue;
-        if (left == 0 && right == 0 && _font_size.x_offset < 0) left = right = -_font_size.x_offset;
-        right = left + std::max((int32_t)_font_size.x_advance, _font_size.x_size + _font_size.x_offset);
-        left += _font_size.x_advance;
-      } while (*(++string));
-
-      return right * _textsize_x;
-    }
-
-  #if defined (ARDUINO)
-    inline int32_t textWidth(const String& string) { return textWidth(string.c_str()); }
-
-    inline int_fast16_t drawString(const String& string, int32_t x, int32_t y) { return draw_string(string.c_str(), x, y, _textdatum); }
-
-    [[deprecated("use setTextDatum() and drawString()")]]
-    inline int_fast16_t drawCentreString(const String& string, int32_t x, int32_t y, uint8_t font) { setFont(fontdata[font]); return draw_string(string.c_str(), x, y, textdatum_t::top_center); }
-
-    [[deprecated("use setTextDatum() and drawString()")]]
-    inline int_fast16_t drawCenterString(const String& string, int32_t x, int32_t y, uint8_t font) { setFont(fontdata[font]); return draw_string(string.c_str(), x, y, textdatum_t::top_center); }
-
-    [[deprecated("use setTextDatum() and drawString()")]]
-    inline int_fast16_t drawRightString( const String& string, int32_t x, int32_t y, uint8_t font) { setFont(fontdata[font]); return draw_string(string.c_str(), x, y, textdatum_t::top_right); }
-
-    inline int_fast16_t drawString(const String& string, int32_t x, int32_t y, uint8_t font) { setFont(fontdata[font]); return draw_string(string.c_str(), x, y, _textdatum); }
-
-  #endif
-    inline int_fast16_t drawString(const char *string, int32_t x, int32_t y) { return draw_string(string, x, y, _textdatum); }
-
-    [[deprecated("use setTextDatum() and drawString()")]]
-    inline int_fast16_t drawCentreString(const char *string, int32_t x, int32_t y, uint8_t font) { setFont(fontdata[font]); return draw_string(string, x, y, textdatum_t::top_center); }
-
-    [[deprecated("use setTextDatum() and drawString()")]]
-    inline int_fast16_t drawCenterString(const char *string, int32_t x, int32_t y, uint8_t font) { setFont(fontdata[font]); return draw_string(string, x, y, textdatum_t::top_center); }
-
-    [[deprecated("use setTextDatum() and drawString()")]]
-    inline int_fast16_t drawRightString( const char *string, int32_t x, int32_t y, uint8_t font) { setFont(fontdata[font]); return draw_string(string, x, y, textdatum_t::top_right); }
-
-    inline int_fast16_t drawString(const char *string, int32_t x, int32_t y, uint8_t font) { setFont(fontdata[font]); return draw_string(string, x, y, _textdatum); }
-
-    inline int_fast16_t drawNumber(long long_num, int32_t poX, int32_t poY, uint8_t font) { setFont(fontdata[font]); return drawNumber(long_num, poX, poY); }
-
-    inline int_fast16_t drawFloat(float floatNumber, uint8_t dp, int32_t poX, int32_t poY, uint8_t font) { setFont(fontdata[font]); return drawFloat(floatNumber, dp, poX, poY); }
-
-    int_fast16_t drawNumber(long long_num, int32_t poX, int32_t poY)
-    {
-      constexpr size_t len = 8 * sizeof(long) + 1;
-      char buf[len];
-      return drawString(numberToStr(long_num, buf, len, 10), poX, poY);
-    }
-
-    int_fast16_t drawFloat(float floatNumber, uint8_t dp, int32_t poX, int32_t poY)
-    {
-      size_t len = 14 + dp;
-      char buf[len];
-      return drawString(floatToStr(floatNumber, buf, len, dp), poX, poY);
-    }
-
-    uint16_t decodeUTF8(uint8_t c)
-    {
-      // 7 bit Unicode Code Point
-      if (!(c & 0x80)) {
-        _decoderState = utf8_decode_state_t::utf8_state0;
-        return c;
-      }
-
-      if (_decoderState == utf8_decode_state_t::utf8_state0)
-      {
-        // 11 bit Unicode Code Point
-        if ((c & 0xE0) == 0xC0)
-        {
-          _unicode_buffer = ((c & 0x1F)<<6);
-          _decoderState = utf8_decode_state_t::utf8_state1;
-          return 0;
-        }
-
-        // 16 bit Unicode Code Point
-        if ((c & 0xF0) == 0xE0)
-        {
-          _unicode_buffer = ((c & 0x0F)<<12);
-          _decoderState = utf8_decode_state_t::utf8_state2;
-          return 0;
-        }
-        // 21 bit Unicode  Code Point not supported so fall-back to extended ASCII
-        //if ((c & 0xF8) == 0xF0) return (uint16_t)c;
-      }
-      else
-      {
-        if (_decoderState == utf8_decode_state_t::utf8_state2)
-        {
-          _unicode_buffer |= ((c & 0x3F)<<6);
-          _decoderState = utf8_decode_state_t::utf8_state1;
-          return 0;
-        }
-        _unicode_buffer |= (c & 0x3F);
-        _decoderState = utf8_decode_state_t::utf8_state0;
-        return _unicode_buffer;
-      }
-
-      _decoderState = utf8_decode_state_t::utf8_state0;
-
-      return (uint16_t)c; // fall-back to extended ASCII
-    }
-
-    template<typename T>
-    void setTextColor(T color) {
-      if (hasPalette()) {
-        _text_fore_rgb888 = _text_back_rgb888 = color;
-      } else {
-        _text_fore_rgb888 = _text_back_rgb888 = convert_to_rgb888(color);
-      }
-    }
-    template<typename T1, typename T2>
-    void setTextColor(T1 fgcolor, T2 bgcolor) {
-      if (hasPalette()) {
-        _text_fore_rgb888 = fgcolor;
-        _text_back_rgb888 = bgcolor;
-      } else {
-        _text_fore_rgb888 = convert_to_rgb888(fgcolor);
-        _text_back_rgb888 = convert_to_rgb888(bgcolor);
-      }
-    }
-
-    inline int_fast16_t drawChar(uint16_t uniCode, int32_t x, int32_t y) { _filled_x = 0; return (fpDrawChar)(this, x, y, uniCode, _text_fore_rgb888, _text_back_rgb888, _textsize_x, _textsize_y, _font); }
-
-    int_fast16_t drawChar(uint16_t uniCode, int32_t x, int32_t y, uint8_t font) {
-      if (_font == fontdata[font]) return drawChar(uniCode, x, y);
-      _filled_x = 0;
-      switch (fontdata[font]->getType()) {
-      default:
-      case font_type_t::ft_glcd: return drawCharGLCD(this, x, y, uniCode, _text_fore_rgb888, _text_back_rgb888, _textsize_x, _textsize_y, fontdata[font]);
-      case font_type_t::ft_bmp:  return drawCharBMP( this, x, y, uniCode, _text_fore_rgb888, _text_back_rgb888, _textsize_x, _textsize_y, fontdata[font]);
-      case font_type_t::ft_rle:  return drawCharRLE( this, x, y, uniCode, _text_fore_rgb888, _text_back_rgb888, _textsize_x, _textsize_y, fontdata[font]);
-      case font_type_t::ft_bdf:  return drawCharBDF( this, x, y, uniCode, _text_fore_rgb888, _text_back_rgb888, _textsize_x, _textsize_y, fontdata[font]);
-      }
-    }
-
-    template<typename T>
-    inline int_fast16_t drawChar(int32_t x, int32_t y, uint16_t uniCode, T color, T bg, int_fast8_t size) { _filled_x = 0; return (fpDrawChar)(this, x, y, uniCode, convert_to_rgb888(color), convert_to_rgb888(bg), size, size, _font); }
-    template<typename T>
-    inline int_fast16_t drawChar(int32_t x, int32_t y, uint16_t uniCode, T color, T bg, int_fast8_t size_x, int_fast8_t size_y) { _filled_x = 0; return (fpDrawChar)(this, x, y, uniCode, convert_to_rgb888(color), convert_to_rgb888(bg), size_x, size_y, _font); }
-
-    int16_t getCursorX(void) const { return _cursor_x; }
-    int16_t getCursorY(void) const { return _cursor_y; }
-    void setCursor( int16_t x, int16_t y)               { _filled_x = 0; _cursor_x = x; _cursor_y = y; }
-    void setCursor( int16_t x, int16_t y, uint8_t font) { _filled_x = 0; _cursor_x = x; _cursor_y = y; _font = fontdata[font]; }
-    void setTextSize(uint8_t s) { setTextSize(s,s); }
-    void setTextSize(uint8_t sx, uint8_t sy) { _textsize_x = (sx > 0) ? sx : 1; _textsize_y = (sy > 0) ? sy : 1; }
-    int16_t getTextSizeX(void) const { return _textsize_x; }
-    int16_t getTextSizeY(void) const { return _textsize_y; }
-    textdatum_t getTextDatum(void) const { return _textdatum; }
-    int16_t fontHeight(void) const { return _font_size.y_size * _textsize_y; }
-    int16_t fontHeight(uint8_t font) const { return pgm_read_byte( &((const BaseFont*)fontdata[font])->height ) * _textsize_y; }
-
-    void setTextDatum(uint8_t datum) { _textdatum = (textdatum_t)datum; }
-    void setTextDatum(textdatum_t datum) { _textdatum = datum; }
-    void setTextPadding(uint16_t padding_x) { _padding_x = padding_x; }
-    void setTextWrap( bool wrapX, bool wrapY = false) { _textwrap_x = wrapX; _textwrap_y = wrapY; }
-    void setTextScroll(bool scroll) { _textscroll = scroll; if (_cursor_x < _sx) { _cursor_x = _sx; } if (_cursor_y < _sy) { _cursor_y = _sy; } }
-
-
-#ifdef __EFONT_FONT_DATA_H__
-    [[deprecated("use setFont(&fonts::efont)")]]
-    void setTextEFont() { setFont(&fonts::efont); }
-#endif
-
-    uint8_t getTextFont(void) const {
-      size_t ie = sizeof(fontdata) / sizeof(fontdata[0]);
-      for (size_t i = 0; i < ie; ++i)
-        if (fontdata[i] == _font) return i;
-      return 0;
-    }
-
-    [[deprecated("use setFont(&fonts::font0)")]]
-    void setTextFont(int f) {
-      if (f == 1 && _font && _font->getType() == ft_gfx) return;
-
-      setFont(fontdata[f]);
-    }
-
-    [[deprecated("use setFont(&fonts::font0)")]]
-    void setTextFont(const IFont* font = nullptr) { setFont(font); }
-
-    [[deprecated("use setFont(&fonts::font0)")]]
-    void setFreeFont(const IFont* font = nullptr) { setFont(font); }
-
-    void unloadFont(void) {
-      if (_dynamic_font) {
-        //_dynamic_font->unloadFont();
-        delete _dynamic_font;
-        _dynamic_font = nullptr;
-      }
-      setFont(&fonts::font0);
-    }
-
-    void setFont(const IFont* font) {
-      if (_dynamic_font) {
-        delete _dynamic_font;
-        _dynamic_font = nullptr;
-      }
-      if (font == nullptr) font = &fonts::font0;
-      _font = font;
-      _filled_x = 0;
-      //_decoderState = utf8_decode_state_t::utf8_state0;
-
-      font->getDefaultMetric(&_font_size);
-
-      switch (font->getType()) {
-      default:
-      case font_type_t::ft_glcd: fpDrawChar = drawCharGLCD;  break;
-      case font_type_t::ft_bmp:  fpDrawChar = drawCharBMP;   break;
-      case font_type_t::ft_rle:  fpDrawChar = drawCharRLE;   break;
-      case font_type_t::ft_bdf:  fpDrawChar = drawCharBDF;   break;
-      case font_type_t::ft_gfx:  fpDrawChar = drawCharGFXFF; break;
-      }
-    }
-
-    void cp437(bool enable = true) { _cp437 = enable; }  // AdafruitGFX compatible.
-
-    void setAttribute(uint8_t attr_id, uint8_t param) { setAttribute((attribute_t)attr_id, param); }
-    void setAttribute(attribute_t attr_id, uint8_t param) {
-      switch (attr_id) {
-        case cp437_switch:
-            _cp437 = param;
-            break;
-        case utf8_switch:
-            _utf8  = param;
-            _decoderState = utf8_decode_state_t::utf8_state0;
-            break;
-        default: break;
-      }
-    }
-
-    uint8_t getAttribute(uint8_t attr_id) { return getAttribute((attribute_t)attr_id); }
-    uint8_t getAttribute(attribute_t attr_id) {
-      switch (attr_id) {
-        case cp437_switch: return _cp437;
-        case utf8_switch: return _utf8;
-        default: return 0;
-      }
-    }
 
 //----------------------------------------------------------------------------
 //----------------------------------------------------------------------------
@@ -1618,9 +1212,6 @@ namespace lgfx
     int32_t  _sx, _sy, _sw, _sh; // for scroll zone
 
     int32_t _clip_l = 0, _clip_r = -1, _clip_t = 0, _clip_b = -1; // clip rect
-    int32_t _cursor_x = 0, _cursor_y = 0, _filled_x = 0;
-    uint32_t _text_fore_rgb888 = 0xFFFFFFU;
-    uint32_t _text_back_rgb888 = 0;
     uint32_t _scolor;  // gap fill colour for scroll zone
     raw_color_t _color = 0xFFFFFFU;
 
@@ -1630,38 +1221,11 @@ namespace lgfx
     bgr888_t* _palette = nullptr; // for sprite palette mode.
     uint32_t _palette_count = 0;
 
-    int16_t _textsize_x = 1;
-    int16_t _textsize_y = 1;
     int16_t _xpivot;   // x pivot point coordinate
     int16_t _ypivot;   // x pivot point coordinate
-    uint16_t _unicode_buffer = 0;   // Unicode code-point buffer
-    //uint8_t _textfont = 1;
-
-    const IFont* _font = &fonts::font0;
-    IFont* _dynamic_font = nullptr;  // run-time generated font
-
-    textdatum_t _textdatum;
-    int16_t _padding_x = 0;
 
     bool _spi_shared = true;
     bool _swapBytes = false;
-    bool _textwrap_x = true;
-    bool _textwrap_y = false;
-    bool _textscroll = false;
-    bool _cp437      = false;
-    bool _utf8       = true;
-
-    enum utf8_decode_state_t
-    { utf8_state0 = 0
-    , utf8_state1 = 1
-    , utf8_state2 = 2
-    };
-    utf8_decode_state_t _decoderState = utf8_state0;   // UTF8 decoder state
-
-    font_size_t _font_size = { 6, 6, 0, 8, 8, 0, 6 };
-//  font_size_t _font_size_x = { 6, 6, 0 };
-//  font_size_t _font_size_y = { 8, 8, 0 };
-//  int8_t _font_size.baseline = 6;
 
     __attribute__ ((always_inline)) inline static bool _adjust_abs(int32_t& x, int32_t& w) { if (w < 0) { x += w + 1; w = -w; } return !w; }
 
@@ -1760,544 +1324,6 @@ namespace lgfx
     virtual void pushBlock_impl(int32_t len) = 0;
     virtual void setWindow_impl(int32_t xs, int32_t ys, int32_t xe, int32_t ye) = 0;
 
-//----------------------------------------------------------------------------
-// print & text support
-//----------------------------------------------------------------------------
-// Arduino Print.h compatible
-    size_t printNumber(unsigned long n, uint8_t base)
-    {
-      size_t len = 8 * sizeof(long) + 1;
-      char buf[len];
-      return write(numberToStr(n, buf, len, base));
-    }
-
-    size_t printFloat(double number, uint8_t digits)
-    {
-      size_t len = 14 + digits;
-      char buf[len];
-      return write(floatToStr(number, buf, len, digits));
-    }
-
-    char* numberToStr(long n, char* buf, size_t buflen, uint8_t base)
-    {
-      if (n >= 0) return numberToStr((unsigned long) n, buf, buflen, base);
-      auto res = numberToStr(- n, buf, buflen, 10) - 1;
-      res[0] = '-';
-      return res;
-    }
-
-    char* numberToStr(unsigned long n, char* buf, size_t buflen, uint8_t base)
-    {
-      char *str = &buf[buflen - 1];
-
-      *str = '\0';
-
-      if (base < 2) { base = 10; }  // prevent crash if called with base == 1
-      do {
-        unsigned long m = n;
-        n /= base;
-        char c = m - base * n;
-        *--str = c < 10 ? c + '0' : c + 'A' - 10;
-      } while (n);
-
-      return str;
-    }
-
-    char* floatToStr(double number, char* buf, size_t buflen, uint8_t digits)
-    {
-      if (std::isnan(number))    { return strcpy(buf, "nan"); }
-      if (std::isinf(number))    { return strcpy(buf, "inf"); }
-      if (number > 4294967040.0) { return strcpy(buf, "ovf"); } // constant determined empirically
-      if (number <-4294967040.0) { return strcpy(buf, "ovf"); } // constant determined empirically
-
-      char* dst = buf;
-      // Handle negative numbers
-      //bool negative = (number < 0.0);
-      if (number < 0.0) {
-        number = -number;
-        *dst++ = '-';
-      }
-
-      // Round correctly so that print(1.999, 2) prints as "2.00"
-      double rounding = 0.5;
-      for(uint8_t i = 0; i < digits; ++i) {
-        rounding /= 10.0;
-      }
-
-      number += rounding;
-
-      // Extract the integer part of the number and print it
-      unsigned long int_part = (unsigned long) number;
-      double remainder = number - (double) int_part;
-
-      {
-        constexpr size_t len = 14;
-        char numstr[len];
-        auto tmp = numberToStr(int_part, numstr, len, 10);
-        auto slen = strlen(tmp);
-        memcpy(dst, tmp, slen);
-        dst += slen;
-      }
-
-      // Print the decimal point, but only if there are digits beyond
-      if (digits > 0) {
-        dst[0] = '.';
-        ++dst;
-      }
-      // Extract digits from the remainder one at a time
-      while (digits-- > 0) {
-        remainder *= 10.0;
-        unsigned int toPrint = (unsigned int)(remainder);
-        dst[0] = '0' + toPrint;
-        ++dst;
-        remainder -= toPrint;
-      }
-      dst[0] = 0;
-      return buf;
-    }
-
-    int_fast16_t draw_string(const char *string, int32_t x, int32_t y, textdatum_t datum)
-    {
-      int16_t sumX = 0;
-      int32_t cwidth = textWidth(string); // Find the pixel width of the string in the font
-      int32_t cheight = _font_size.y_size * _textsize_y;
-
-      {
-        auto tmp = string;
-        do {
-          uint16_t uniCode = *tmp;
-          if (_utf8) {
-            do {
-              uniCode = decodeUTF8(*tmp); 
-            } while (uniCode < 32 && *++tmp);
-            if (uniCode < 32) break;
-          }
-        //if ((fpUpdateFontSize)(this, uniCode)) {
-          if (_font->updateFontMetric(&_font_size, uniCode)) {
-            if (_font_size.x_offset < 0) sumX = - _font_size.x_offset * _textsize_x;
-            break;
-          }
-        } while (*++tmp);
-      }
-      if (datum & middle_left) {          // vertical: middle
-        y -= cheight >> 1;
-      } else if (datum & bottom_left) {   // vertical: bottom
-        y -= cheight;
-      } else if (datum & baseline_left) { // vertical: baseline
-        y -= _font_size.baseline * _textsize_y;
-      }
-
-      startWrite();
-      int32_t padx = _padding_x;
-      if ((_text_fore_rgb888 != _text_back_rgb888) && (padx > cwidth)) {
-        setColor(_text_back_rgb888);
-        if (datum & top_center) {
-          auto halfcwidth = cwidth >> 1;
-          auto halfpadx = (padx >> 1);
-          writeFillRect(x - halfpadx, y, halfpadx - halfcwidth, cheight);
-          halfcwidth = cwidth - halfcwidth;
-          halfpadx = padx - halfpadx;
-          writeFillRect(x + halfcwidth, y, halfpadx - halfcwidth, cheight);
-        } else if (datum & top_right) {
-          writeFillRect(x - padx, y, padx - cwidth, cheight);
-        } else {
-          writeFillRect(x + cwidth, y, padx - cwidth, cheight);
-        }
-      }
-
-      if (datum & top_center) {           // Horizontal: middle
-        x -= cwidth >> 1;
-      } else if (datum & top_right) {     // Horizontal: right
-        x -= cwidth;
-      }
-
-      y -= _font_size.y_offset * _textsize_y;
-
-      _filled_x = 0;
-      do {
-        uint16_t uniCode = *string;
-        if (_utf8) {
-          do {
-            uniCode = decodeUTF8(*string);
-          } while (uniCode < 32 && *++string);
-          if (uniCode < 32) break;
-        }
-        sumX += (fpDrawChar)(this, x + sumX, y, uniCode, _text_fore_rgb888, _text_back_rgb888, _textsize_x, _textsize_y, _font);
-      } while (*(++string));
-      endWrite();
-
-      return sumX;
-    }
-
-    //bool(*fpUpdateFontSize)(LGFXBase* me, uint16_t uniCode) = updateFontSizeGLCD;
-//    static bool updateFontSizeGLCD(LGFXBase* me, uint16_t uniCode) {
-//      return uniCode < 256;
-//    }
-//    static bool updateFontSizeBMP(LGFXBase* me, uint16_t uniCode) {
-//      if ((uniCode -= 32) >= 96) return false;
-//      auto font = (const GLCDfont*)me->_font;
-//      me->_font_size.x_advance = me->_font_size.x_size = pgm_read_byte( (uint8_t *)pgm_read_dword( & font->widthtbl ) + uniCode );
-//      return true;
-//    }
-//    static bool updateFontSizeBDF(LGFXBase* me, uint16_t uniCode) {
-//      auto fontd = (BDFfont*)(me->_font);
-//      me->_font_size.x_size = me->_font_size.x_advance = (uniCode < 0x0100) ? fontd->halfwidth : fontd->width;
-//      return true;
-//    }
-
-    int_fast16_t (*fpDrawChar)(LGFXBase* me, int32_t x, int32_t y, uint16_t c, uint32_t fore_rgb888, uint32_t back_rgb888, int_fast8_t size_x, int_fast8_t size_y, const IFont* font) = &LGFXBase::drawCharGLCD;
-
-    static int_fast16_t drawCharGLCD(LGFXBase* me, int32_t x, int32_t y, uint16_t c, uint32_t fore_rgb888, uint32_t back_rgb888, int_fast8_t size_x, int_fast8_t size_y, const IFont* font)
-    { // glcd font
-      auto fontdat = (const GLCDfont*)font;
-      if (c > 255) return 0;
-
-      if (!me->_cp437 && (c >= 176)) c++; // Handle 'classic' charset behavior
-
-      const int32_t fontWidth  = me->_font_size.x_size;
-      const int32_t fontHeight = me->_font_size.y_size;
-
-      auto font_addr = fontdat->chartbl + (c * 5);
-      uint32_t colortbl[2] = {me->_write_conv.convert(back_rgb888), me->_write_conv.convert(fore_rgb888)};
-      bool fillbg = (back_rgb888 != fore_rgb888);
-
-      int32_t clip_left   = me->_clip_l;
-      int32_t clip_right  = me->_clip_r;
-      int32_t clip_top    = me->_clip_t;
-      int32_t clip_bottom = me->_clip_b;
-
-      if ((x <= clip_right) && (clip_left < (x + fontWidth * size_x ))
-       && (y <= clip_bottom) && (clip_top < (y + fontHeight * size_y ))) {
-        if (!fillbg || size_y > 1 || x < clip_left || y < clip_top || y + fontHeight > clip_bottom || x + fontWidth * size_x > clip_right) {
-          int32_t xpos = x;
-          me->startWrite();
-
-          int_fast8_t i = 0;
-          do {
-            int_fast16_t ypos = y;
-            uint8_t line = pgm_read_byte(&font_addr[i]);
-            uint8_t flg = (line & 0x01);
-            int_fast8_t j = 1;
-            int_fast8_t jp = 0;
-            do {
-              while (flg == ((line >> j) & 0x01) && ++j < fontHeight);
-              jp = j - jp;
-              if (flg || fillbg) {
-                me->setRawColor(colortbl[flg]);
-                me->writeFillRect(xpos, ypos, size_x, jp * size_y);
-              }
-              ypos += jp * size_y;
-              flg = !flg;
-              jp = j;
-            } while (j < fontHeight);
-            xpos += size_x;
-          } while (++i < fontWidth - 1);
-
-          if (fillbg) {
-            me->setRawColor(colortbl[0]);
-            me->writeFillRect(xpos, y, size_x, fontHeight * size_y); 
-          }
-          me->endWrite();
-        } else {
-          uint8_t col[fontWidth];
-          int_fast8_t i = 0;
-          do {
-            col[i] = pgm_read_byte(&font_addr[i]);
-          } while (++i < 5);
-          col[5] = 0;
-          me->startWrite();
-          me->setAddrWindow(x, y, fontWidth * size_x, fontHeight);
-          uint8_t flg = col[0] & 1;
-          uint32_t len = 0;
-          i = 0;
-          do {
-            int_fast8_t j = 0;
-            do {
-              if (flg != ((col[j] >> i) & 1)) {
-                me->writeRawColor(colortbl[flg], len);
-                len = 0;
-                flg = !flg;
-              }
-              len += size_x;
-            } while (++j < fontWidth);
-          } while (++i < fontHeight);
-          me->writeRawColor(colortbl[0], len);
-          me->endWrite();
-        }
-      }
-      return fontWidth * size_x;
-    }
-
-    static int_fast16_t drawCharBMP(LGFXBase* me, int32_t x, int32_t y, uint16_t uniCode, uint32_t fore_rgb888, uint32_t back_rgb888, int_fast8_t size_x, int_fast8_t size_y, const IFont* font)
-    { // BMP font
-      auto fontdat = (const BMPfont*)font;
-
-      if ((uniCode -= 32) >= 96) return 0;
-      const int_fast8_t fontWidth = pgm_read_byte(fontdat->widthtbl + uniCode);
-      const int_fast8_t fontHeight = pgm_read_byte(&fontdat->height);
-
-      auto font_addr = (const uint8_t*)pgm_read_dword(&((const uint8_t**)fontdat->chartbl)[uniCode]);
-      return draw_char_bmp(me, x, y, fore_rgb888, back_rgb888, size_x, size_y, font_addr, fontWidth, fontHeight, (fontWidth + 6) >> 3, 1);
-    }
-
-    static int_fast16_t drawCharBDF(LGFXBase* me, int32_t x, int32_t y, uint16_t c, uint32_t fore_rgb888, uint32_t back_rgb888, int_fast8_t size_x, int_fast8_t size_y, const IFont* font)
-    {
-      auto fontd = (BDFfont*)font;
-      const int_fast8_t bytesize = (fontd->width + 7) >> 3;
-      const int_fast8_t fontHeight = fontd->height;
-      const int_fast8_t fontWidth = (c < 0x0100) ? fontd->halfwidth : fontd->width;
-      auto it = std::lower_bound(fontd->indextbl, &fontd->indextbl[fontd->indexsize], c);
-      if (*it != c) {
-        if (fore_rgb888 != back_rgb888) {
-          me->fillRect(x, y, fontWidth * size_x, fontHeight * size_y, back_rgb888);
-        }
-        return fontWidth * size_x;
-      }
-      const uint8_t* font_addr = &fontd->chartbl[std::distance(fontd->indextbl, it) * fontHeight * bytesize];
-      return LGFXBase::draw_char_bmp(me, x, y, fore_rgb888, back_rgb888, size_x, size_y, font_addr, fontWidth, fontHeight, bytesize, 0);
-    }
-
-    static int_fast16_t draw_char_bmp(LGFXBase* me, int32_t x, int32_t y, uint32_t fore_rgb888, uint32_t back_rgb888, int_fast8_t size_x, int_fast8_t size_y, const uint8_t* font_addr, int_fast8_t fontWidth, int_fast8_t fontHeight, int_fast8_t w, int_fast8_t margin )
-    {
-      uint32_t colortbl[2] = {me->_write_conv.convert(back_rgb888), me->_write_conv.convert(fore_rgb888)};
-      bool fillbg = (back_rgb888 != fore_rgb888);
-
-      int32_t clip_left   = me->_clip_l;
-      int32_t clip_right  = me->_clip_r;
-      int32_t clip_top    = me->_clip_t;
-      int32_t clip_bottom = me->_clip_b;
-
-      if ((x <= clip_right) && (clip_left < (x + fontWidth * size_x ))
-       && (y <= clip_bottom) && (clip_top < (y + fontHeight * size_y ))) {
-        if (!fillbg || size_y > 1 || x < clip_left || y < clip_top || y + fontHeight > clip_bottom || x + fontWidth * size_x > clip_right) {
-          me->startWrite();
-          if (fillbg) {
-            me->setRawColor(colortbl[0]);
-            if (margin)
-              me->writeFillRect(x + (fontWidth - margin) * size_x, y, size_x, fontHeight * size_y);
-          }
-          int_fast8_t i = 0;
-          do {
-            uint8_t line = pgm_read_byte(font_addr);
-            bool flg = line & 0x80;
-            int_fast8_t len = 1;
-            int_fast8_t j = 1;
-            int_fast8_t je = fontWidth - margin;
-            do {
-              if (j & 7) {
-                line <<= 1;
-              } else {
-                line = pgm_read_byte(&font_addr[j >> 3]);
-              }
-              if (flg != (bool)(line & 0x80)) {
-                if (flg || fillbg) {
-                  me->setRawColor(colortbl[flg]);
-                  me->writeFillRect( x + (j - len) * size_x, y, len * size_x, size_y); 
-                }
-                len = 1;
-                flg = !flg;
-              } else {
-                ++len;
-              }
-            } while (++j < je);
-            if (flg || fillbg) {
-              me->setRawColor(colortbl[flg]);
-              me->writeFillRect( x + (j - len) * size_x, y, len * size_x, size_y); 
-            }
-            y += size_y;
-            font_addr += w;
-          } while (++i < fontHeight);
-          me->endWrite();
-        } else {
-          int_fast8_t len = 0;
-          uint8_t line = 0;
-          bool flg = false;
-          me->startWrite();
-          me->setAddrWindow(x, y, fontWidth * size_x, fontHeight);
-          int_fast8_t i = 0;
-          int_fast8_t je = fontWidth - margin;
-          do {
-            int_fast8_t j = 0;
-            do {
-              if (j & 7) {
-                line <<= 1;
-              } else {
-                line = (j == je) ? 0 : pgm_read_byte(&font_addr[j >> 3]);
-              }
-              if (flg != (bool)(line & 0x80)) {
-                me->writeRawColor(colortbl[flg], len);
-                flg = !flg;
-                len = 0;
-              }
-              len += size_x;
-            } while (++j < fontWidth);
-            font_addr += w;
-          } while (++i < fontHeight);
-          me->writeRawColor(colortbl[flg], len);
-          me->endWrite();
-        }
-      }
-
-      return fontWidth * size_x;
-    }
-
-    static int_fast16_t drawCharRLE(LGFXBase* me, int32_t x, int32_t y, uint16_t code, uint32_t fore_rgb888, uint32_t back_rgb888, int_fast8_t size_x, int_fast8_t size_y, const IFont* font)
-    { // RLE font
-      auto fontdat = (RLEfont*)font;
-      if ((code -= 32) >= 96) return 0;
-
-      const int fontWidth = pgm_read_byte( (uint8_t *)pgm_read_dword( &(fontdat->widthtbl ) ) + code );
-      const int fontHeight = pgm_read_byte( &fontdat->height );
-
-      auto font_addr = (const uint8_t*)pgm_read_dword( (const void*)(pgm_read_dword( &(fontdat->chartbl ) ) + code * sizeof(void *)) );
-
-      uint32_t colortbl[2] = {me->_write_conv.convert(back_rgb888), me->_write_conv.convert(fore_rgb888)};
-      bool fillbg = (back_rgb888 != fore_rgb888);
-
-      int32_t clip_left   = me->_clip_l;
-      int32_t clip_right  = me->_clip_r;
-      int32_t clip_top    = me->_clip_t;
-      int32_t clip_bottom = me->_clip_b;
-
-      if ((x <= clip_right) && (clip_left < (x + fontWidth * size_x ))
-       && (y <= clip_bottom) && (clip_top < (y + fontHeight * size_y ))) {
-        if (!fillbg || size_y > 1 || x < clip_left || y < clip_top || y + fontHeight > clip_bottom || x + fontWidth * size_x > clip_right) {
-          bool flg = false;
-          uint8_t line = 0, i = 0, j = 0;
-          int32_t len;
-          me->startWrite();
-          do {
-            line = pgm_read_byte(font_addr++);
-            flg = line & 0x80;
-            line = (line & 0x7F)+1;
-            do {
-              len = (j + line > fontWidth) ? fontWidth - j : line;
-              line -= len;
-              if (fillbg || flg) {
-                me->setRawColor(colortbl[flg]);
-                me->writeFillRect( x + j * size_x, y + (i * size_y), len * size_x, size_y);
-              }
-              j += len;
-              if (j == fontWidth) {
-                j = 0;
-                i++;
-              }
-            } while (line);
-          } while (i < fontHeight);
-          me->endWrite();
-        } else {
-          uint32_t line = 0;
-          me->startWrite();
-          me->setAddrWindow(x, y, fontWidth * size_x, fontHeight);
-          uint32_t len = fontWidth * size_x * fontHeight;
-          do {
-            line = pgm_read_byte(font_addr++);
-            bool flg = line & 0x80;
-            line = ((line & 0x7F) + 1) * size_x;
-            me->writeRawColor(colortbl[flg], line);
-          } while (len -= line);
-          me->endWrite();
-        }
-      }
-
-      return fontWidth * size_x;
-    }
-
-//    static bool updateFontSizeGFXFF(LGFXBase* me, uint16_t uniCode) {
-//      auto glyph = ((const GFXfont*)me->_font)->getGlyph(uniCode);
-//      if (!glyph) return false;
-//      me->_font_size.x_offset  = (int8_t)pgm_read_byte(&glyph->xOffset);
-//      me->_font_size.x_size    = pgm_read_byte(&glyph->width);
-//      me->_font_size.x_advance = pgm_read_byte(&glyph->xAdvance);
-//      return true;
-//    }
-
-    static int_fast16_t drawCharGFXFF(LGFXBase* me, int32_t x, int32_t y, uint16_t uniCode, uint32_t fore_rgb888, uint32_t back_rgb888, int_fast8_t size_x, int_fast8_t size_y, const IFont* fontdat)
-    {
-      auto gfxFont = (const GFXfont*)me->_font;
-      auto glyph = gfxFont->getGlyph(uniCode);
-      if (!glyph) return 0;
-
-      int32_t w = pgm_read_byte(&glyph->width),
-              h = pgm_read_byte(&glyph->height);
-
-      int32_t xAdvance = (int32_t)size_x * (int8_t)pgm_read_byte(&glyph->xAdvance);
-      int32_t xoffset  = (int32_t)size_x * (int8_t)pgm_read_byte(&glyph->xOffset);
-      int32_t yoffset  = (int32_t)size_y * (int8_t)pgm_read_byte(&glyph->yOffset);
-
-      me->startWrite();
-      uint32_t colortbl[2] = {me->_write_conv.convert(back_rgb888), me->_write_conv.convert(fore_rgb888)};
-      bool fillbg = (back_rgb888 != fore_rgb888);
-      int32_t left  = 0;
-      int32_t right = 0;
-      if (fillbg) {
-        left  = std::max(me->_filled_x, x + (xoffset < 0 ? xoffset : 0));
-        right = x + std::max((int32_t)(w * size_x + xoffset), (int32_t)(xAdvance));
-      }
-
-      x += xoffset;
-      y += yoffset;
-
-      int32_t clip_left   = me->_clip_l;
-      int32_t clip_right  = me->_clip_r;
-      int32_t clip_top    = me->_clip_t;
-      int32_t clip_bottom = me->_clip_b;
-
-      if ((x <= clip_right) && (clip_left < (x + w * size_x ))
-       && (y <= clip_bottom) && (clip_top < (y + h * size_y ))) {
-
-        if (right > left) {
-          me->setRawColor(colortbl[0]);
-          int tmp = yoffset - (me->_font_size.y_offset * size_y);
-          if (tmp > 0)
-            me->writeFillRect(left, y - yoffset + me->_font_size.y_offset * size_y, right - left, tmp);
-
-          tmp = (me->_font_size.y_offset + me->_font_size.y_size - h) * size_y - yoffset;
-          if (tmp > 0)
-            me->writeFillRect(left, y + h * size_y, right - left, tmp);
-        }
-
-        uint8_t  *bitmap = (uint8_t *)pgm_read_dword(&gfxFont->bitmap)
-                         + pgm_read_dword(&glyph->bitmapOffset);
-        uint8_t bits=0, bit=0;
-
-        me->setRawColor(colortbl[1]);
-        while (h--) {
-          if (right > left) {
-            me->setRawColor(colortbl[0]);
-            me->writeFillRect(left, y, right - left, size_y);
-            me->setRawColor(colortbl[1]);
-          }
-
-          int32_t len = 0;
-          int32_t i = 0;
-          for (i = 0; i < w; i++) {
-            if (bit == 0) {
-              bit  = 0x80;
-              bits = pgm_read_byte(bitmap++);
-            }
-            if (bits & bit) len++;
-            else if (len) {
-              me->writeFillRect(x + (i-len) * size_x, y, size_x * len, size_y);
-              len=0;
-            }
-            bit >>= 1;
-          }
-          if (len) {
-            me->writeFillRect(x + (i-len) * size_x, y, size_x * len, size_y);
-          }
-          y += size_y;
-        }
-      } else {
-        if (right > left) {
-          me->setRawColor(colortbl[0]);
-          me->writeFillRect(left, y - yoffset + me->_font_size.y_offset * size_y, right - left, (me->_font_size.y_size) * size_y);
-        }
-      }
-      me->_filled_x = right;
-      me->endWrite();
-      return xAdvance;
-    }
 
     static void tmpBeginTransaction(void* lgfx) {
       auto me = (LGFXBase*)lgfx;
@@ -2307,1139 +1333,7 @@ namespace lgfx
       auto me = (LGFXBase*)lgfx;
       if (me->_transaction_count) me->endTransaction();
     }
-
   };
-
-//----------------------------------------------------------------------------
-
-  struct VLWfont : public IFont {
-    uint16_t gCount;     // Total number of characters
-    uint16_t yAdvance;   // Line advance
-    uint16_t spaceWidth; // Width of a space character
-    int16_t  ascent;     // Height of top of 'd' above baseline, other characters may be taller
-    int16_t  descent;    // Offset to bottom of 'p', other characters may have a larger descent
-    uint16_t maxAscent;  // Maximum ascent found in font
-    uint16_t maxDescent; // Maximum descent found in font
-
-    // These are for the metrics for each individual glyph (so we don't need to seek this in file and waste time)
-    uint16_t* gUnicode  = nullptr;  //UTF-16 code, the codes are searched so do not need to be sequential
-    uint8_t*  gWidth    = nullptr;  //cwidth
-    uint8_t*  gxAdvance = nullptr;  //setWidth
-    int8_t*   gdX       = nullptr;  //leftExtent
-    uint32_t* gBitmap   = nullptr;  //file pointer to greyscale bitmap
-
-    DataWrapper* _fontData;
-    bool _fontLoaded = false; // Flags when a anti-aliased font is loaded
-
-    font_type_t getType(void) const override { return ft_vlw;  } 
-
-    void getDefaultMetric(font_size_t *sizedata) const override {
-      sizedata->x_offset  = 0;
-      sizedata->y_offset  = 0;
-      sizedata->baseline  = maxAscent;
-      sizedata->y_advance = yAdvance;
-      sizedata->y_size    = yAdvance;
-    }
-
-    virtual ~VLWfont() {
-      unloadFont();
-    }
-
-    bool unloadFont(void) override {
-      _fontLoaded = false;
-      if (_fontData) { _fontData->close();   _fontData = nullptr; }
-      if (gUnicode)  { heap_free(gUnicode);  gUnicode  = nullptr; }
-      if (gWidth)    { heap_free(gWidth);    gWidth    = nullptr; }
-      if (gxAdvance) { heap_free(gxAdvance); gxAdvance = nullptr; }
-      if (gdX)       { heap_free(gdX);       gdX       = nullptr; }
-      if (gBitmap)   { heap_free(gBitmap);   gBitmap   = nullptr; }
-      return true;
-    }
-
-    bool getUnicodeIndex(uint16_t unicode, uint16_t *index) const
-    {
-      auto poi = std::lower_bound(gUnicode, &gUnicode[gCount], unicode);
-      *index = std::distance(gUnicode, poi);
-      return (*poi == unicode);
-    }
-
-    bool updateFontMetric(font_size_t *sizedata, uint16_t uniCode) const override {
-      uint16_t gNum = 0;
-      if (getUnicodeIndex(uniCode, &gNum)) {
-        if (gWidth && gxAdvance && gdX[gNum]) {
-          sizedata->x_size    = gWidth[gNum];
-          sizedata->x_advance = gxAdvance[gNum];
-          sizedata->x_offset  = gdX[gNum];
-        } else {
-          auto file = _fontData;
-
-          file->preRead();
-          //if (file->need_transaction && me->_transaction_count) me->endTransaction();
-
-          file->seek(28 + gNum * 28);  // headerPtr
-          uint32_t buffer[6];
-          file->read((uint8_t*)buffer, 24);
-          sizedata->x_size    = __builtin_bswap32(buffer[1]); // Width of glyph
-          sizedata->x_advance = __builtin_bswap32(buffer[2]); // xAdvance - to move x cursor
-          sizedata->x_offset  = (int32_t)((int8_t)__builtin_bswap32(buffer[4])); // x delta from cursor
-
-          file->postRead();
-          //if (file->need_transaction && me->_transaction_count) me->beginTransaction();
-        }
-        return true;
-      }
-      return false;
-    }
-
-
-    bool loadFont(DataWrapper* data) {
-      _fontData = data;
-      {
-        uint32_t buf[6];
-        data->read((uint8_t*)buf, 6 * 4); // 24 Byte read
-
-        gCount   = __builtin_bswap32(buf[0]); // glyph count in file
-                 //__builtin_bswap32(buf[1]); // vlw encoder version - discard
-        yAdvance = __builtin_bswap32(buf[2]); // Font size in points, not pixels
-                 //__builtin_bswap32(buf[3]); // discard
-        ascent   = __builtin_bswap32(buf[4]); // top of "d"
-        descent  = __builtin_bswap32(buf[5]); // bottom of "p"
-      }
-
-      // These next gFont values might be updated when the Metrics are fetched
-      maxAscent  = ascent;   // Determined from metrics
-      maxDescent = descent;  // Determined from metrics
-      yAdvance   = std::max((int)yAdvance, ascent + descent);
-      spaceWidth = yAdvance * 2 / 7;  // Guess at space width
-
-//ESP_LOGI("LGFX", "ascent:%d  descent:%d", gFont.ascent, gFont.descent);
-
-      if (!gCount) return false;
-
-//ESP_LOGI("LGFX", "font count:%d", gCount);
-
-      uint32_t bitmapPtr = 24 + (uint32_t)gCount * 28;
-
-      gBitmap   = (uint32_t*)heap_alloc_psram( gCount * 4); // seek pointer to glyph bitmap in the file
-      gUnicode  = (uint16_t*)heap_alloc_psram( gCount * 2); // Unicode 16 bit Basic Multilingual Plane (0-FFFF)
-      gWidth    =  (uint8_t*)heap_alloc_psram( gCount );    // Width of glyph
-      gxAdvance =  (uint8_t*)heap_alloc_psram( gCount );    // xAdvance - to move x cursor
-      gdX       =   (int8_t*)heap_alloc_psram( gCount );    // offset for bitmap left edge relative to cursor X
-
-      if (!gUnicode
-       || !gBitmap
-       || !gWidth
-       || !gxAdvance
-       || !gdX) {
-//ESP_LOGE("LGFX", "can not alloc font table");
-        return false;
-      }
-
-      _fontLoaded = true;
-
-      size_t gNum = 0;
-      _fontData->seek(24);  // headerPtr
-      uint32_t buffer[7];
-      do {
-        _fontData->read((uint8_t*)buffer, 7 * 4); // 28 Byte read
-        uint16_t unicode = __builtin_bswap32(buffer[0]); // Unicode code point value
-        uint32_t w = (uint8_t)__builtin_bswap32(buffer[2]); // Width of glyph
-        if (gUnicode)   gUnicode[gNum]  = unicode;
-        if (gWidth)     gWidth[gNum]    = w;
-        if (gxAdvance)  gxAdvance[gNum] = (uint8_t)__builtin_bswap32(buffer[3]); // xAdvance - to move x cursor
-        if (gdX)        gdX[gNum]       =  (int8_t)__builtin_bswap32(buffer[5]); // x delta from cursor
-
-        uint16_t height = __builtin_bswap32(buffer[1]); // Height of glyph
-        if ((unicode > 0xFF) || ((unicode > 0x20) && (unicode < 0xA0) && (unicode != 0x7F))) {
-          int16_t dY =  (int16_t)__builtin_bswap32(buffer[4]); // y delta from baseline
-//ESP_LOGI("LGFX", "unicode:%x  dY:%d", unicode, dY);
-          if (maxAscent < dY) {
-            maxAscent = dY;
-          }
-          if (maxDescent < (height - dY)) {
-//ESP_LOGI("LGFX", "maxDescent:%d", maxDescent);
-            maxDescent = height - dY;
-          }
-        }
-
-        if (gBitmap)  gBitmap[gNum] = bitmapPtr;
-        bitmapPtr += w * height;
-      } while (++gNum < gCount);
-
-      yAdvance = maxAscent + maxDescent;
-
-//ESP_LOGI("LGFX", "maxDescent:%d", maxDescent);
-      return true;
-    }
-  };
-
-//----------------------------------------------------------------------------
-
-  template <class Base>
-  class LGFX_VLWFont_Support : public Base {
-  public:
-
-#if defined (ARDUINO) && defined (FS_H)
-
-    void loadFont(const char *path, fs::FS &fs) {
-      this->unloadFont();
-
-      _fontFile.setFS(fs);
-
-      loadFont(path);
-    }
-
-#endif
-
-    void loadFont(const uint8_t* array) {
-      this->unloadFont();
-      _fontPointer.set(array);
-      auto font = new VLWfont();
-      this->_dynamic_font = font;
-      if (font->loadFont(&_fontPointer)) {
-        this->_font = font;
-        this->fpDrawChar = drawCharVLW;
-        //this->fpUpdateFontSize = updateFontSizeVLW;
-        this->_font->getDefaultMetric(&this->_font_size);
-      } else {
-        this->unloadFont();
-      }
-    }
-
-    void loadFont(const char *path) {
-      this->unloadFont();
-
-      _fontFile.need_transaction &= this->isSPIShared();
-
-      if (_fontFile.need_transaction) {
-        _fontFile.parent = this;
-        _fontFile.fp_pre_read  = &LGFXBase::tmpEndTransaction;
-        _fontFile.fp_post_read = &LGFXBase::tmpBeginTransaction;
-        this->endTransaction();
-      } else {
-        _fontFile.parent = nullptr;
-        _fontFile.fp_pre_read  = nullptr;
-        _fontFile.fp_post_read = nullptr;
-      }
-      bool result = _fontFile.open(path, "rb");
-      if (!result) {
-        std::string filename = "/";
-        filename += path;
-        filename += ".vlw";
-        result = _fontFile.open(filename.c_str(), "rb");
-      }
-      if (result) {
-        auto font = new VLWfont();
-        this->_dynamic_font = font;
-        if (font->loadFont(&_fontPointer)) {
-          this->_font = font;
-          this->fpDrawChar = drawCharVLW;
-          //this->fpUpdateFontSize = updateFontSizeVLW;
-          this->_font->getDefaultMetric(&this->_font_size);
-        } else {
-          this->unloadFont();
-        }
-      } else {
-        this->unloadFont();
-      }
-
-      if (_fontFile.need_transaction && this->_transaction_count) { this->beginTransaction(); }
-    }
-
-    void showFont(uint32_t td)
-    {
-      auto font = (const VLWfont*)this->_font;
-      if(!font->_fontLoaded) return;
-
-      int16_t x = this->width();
-      int16_t y = this->height();
-      uint32_t timeDelay = 0;    // No delay before first page
-
-      this->fillScreen(this->_text_back_rgb888);
-
-      for (uint16_t i = 0; i < font->gCount; i++)
-      {
-        // Check if this will need a new screen
-        if (x + font->gdX[i] + font->gWidth[i] >= this->width())  {
-          x = - font->gdX[i];
-
-          y += font->yAdvance;
-          if (y + font->maxAscent + font->descent >= this->height()) {
-            x = - font->gdX[i];
-            y = 0;
-            delay(timeDelay);
-            timeDelay = td;
-            this->fillScreen(this->_text_back_rgb888);
-          }
-        }
-
-        this->drawChar(font->gUnicode[i], x, y);
-        x += font->gxAdvance[i];
-        //yield();
-      }
-
-      delay(timeDelay);
-      this->fillScreen(this->_text_back_rgb888);
-      //fontFile.close();
-    }
-
-  protected:
-
-    FileWrapper _fontFile;
-    PointerWrapper _fontPointer;
-
-    static int_fast16_t drawCharVLW(LGFXBase* lgfxbase, int32_t x, int32_t y, uint16_t code, uint32_t fore_rgb888, uint32_t back_rgb888, int_fast8_t size_x, int_fast8_t size_y, const IFont*)
-    {
-      auto me = (LGFX_VLWFont_Support*)lgfxbase;
-      auto font = (const VLWfont*)me->_font;
-
-      uint16_t gNum = 0;
-      if (!font->getUnicodeIndex(code, &gNum)) {
-        return 0;
-      }
-
-      auto file = font->_fontData;
-
-      if (file->need_transaction && me->_transaction_count) me->endTransaction();
-
-      file->seek(28 + gNum * 28);  // headerPtr
-      uint32_t buffer[6];
-      file->read((uint8_t*)buffer, 24);
-      uint32_t h        = __builtin_bswap32(buffer[0]); // Height of glyph
-      uint32_t w        = __builtin_bswap32(buffer[1]); // Width of glyph
-      uint32_t xAdvance = __builtin_bswap32(buffer[2]) * size_x; // xAdvance - to move x cursor
-      int32_t xoffset   = (int32_t)((int8_t)__builtin_bswap32(buffer[4])) * size_x; // x delta from cursor
-      int32_t dY        = (int16_t)__builtin_bswap32(buffer[3]); // y delta from baseline
-      int32_t yoffset = ((int32_t)font->maxAscent - dY) * (int32_t)size_y;
-
-      uint8_t pbuffer[w * h];
-      uint8_t* pixel = pbuffer;
-
-      file->seek(font->gBitmap[gNum]);  // headerPtr
-      file->read(pixel, w * h);
-
-      if (file->need_transaction && me->_transaction_count) me->beginTransaction();
-      me->startWrite();
-
-      uint32_t colortbl[2] = {me->_write_conv.convert(back_rgb888), me->_write_conv.convert(fore_rgb888)};
-      bool fillbg = (back_rgb888 != fore_rgb888);
-      int32_t left  = 0;
-      int32_t right = 0;
-      if (fillbg) {
-        left  = std::max(me->_filled_x, x + (xoffset < 0 ? xoffset : 0));
-        right = x + std::max((int32_t)(w * size_x + xoffset), (int32_t)(xAdvance));
-      }
-      me->_filled_x = right;
-
-      y += yoffset;
-      x += xoffset;
-      int32_t l = 0;
-      int32_t bx = x;
-      int32_t bw = w * size_x;
-      int32_t clip_left = me->_clip_l;
-      if (x < clip_left) { l = -((x - clip_left) / size_x); bw += (x - clip_left); bx = clip_left; }
-      int32_t clip_right = me->_clip_r + 1;
-      if (bw > clip_right - bx) bw = clip_right - bx;
-      if (bw > 0 && (y <= me->_clip_b) && (me->_clip_t < (y + h * size_y))) {
-        int32_t fore_r = ((fore_rgb888>>16)&0xFF);
-        int32_t fore_g = ((fore_rgb888>> 8)&0xFF);
-        int32_t fore_b = ((fore_rgb888)    &0xFF);
-
-        if (fillbg) { // fill background mode
-          if (right > left) {
-            me->setRawColor(colortbl[0]);
-            int tmp = yoffset - (me->_font_size.y_offset * size_y);
-            if (tmp > 0)
-              me->writeFillRect(left, y - yoffset + me->_font_size.y_offset * size_y, right - left, tmp);
-
-            tmp = (me->_font_size.y_offset + me->_font_size.y_size - h) * size_y - yoffset;
-            if (tmp > 0)
-              me->writeFillRect(left, y + h * size_y, right - left, tmp);
-          }
-
-          int32_t back_r = ((back_rgb888>>16)&0xFF);
-          int32_t back_g = ((back_rgb888>> 8)&0xFF);
-          int32_t back_b = ((back_rgb888)    &0xFF);
-          int32_t r = (clip_right - x + size_x - 1) / size_x;
-          if (r > w) r = w;
-          do {
-            if (right > left) {
-              me->setRawColor(colortbl[0]);
-              me->writeFillRect(left, y, right - left, size_y);
-            }
-            int32_t i = l;
-            do {
-              while (pixel[i] != 0xFF) {
-                if (pixel[i] != 0) {
-                  int32_t p = 1 + (uint32_t)pixel[i];
-                  me->setColor(color888( ( fore_r * p + back_r * (257 - p)) >> 8
-                                       , ( fore_g * p + back_g * (257 - p)) >> 8
-                                       , ( fore_b * p + back_b * (257 - p)) >> 8 ));
-                  me->writeFillRect(i * size_x + x, y, size_x, size_y);
-                }
-                if (++i == r) break;
-              }
-              if (i == r) break;
-              int32_t dl = 1;
-              while (i + dl != r && pixel[i + dl] == 0xFF) { ++dl; }
-              me->setRawColor(colortbl[1]);
-              me->writeFillRect(x + i * size_x, y, dl * size_x, size_y);
-              i += dl;
-            } while (i != r);
-            pixel += w;
-            y += size_y;
-          } while (--h);
-
-        } else { // alpha blend mode
-
-          int32_t xshift = (bx - x) % size_x;
-          bgr888_t buf[bw * size_y];
-          pixelcopy_t p(buf, me->_write_conv.depth, rgb888_3Byte, me->hasPalette());
-          do {
-            int32_t by = y;
-            int32_t bh = size_y;
-            if (by < 0) { bh += by; by = 0; }
-            if (bh > 0) {
-              me->readRectRGB(bx, by, bw, bh, (uint8_t*)buf);
-              for (int32_t sx = 0; sx < bw; sx++) {
-                int32_t p = 1 + pixel[left + (sx+xshift) / size_x];
-                int32_t sy = 0;
-                do {
-                  auto bgr = &buf[sx + sy * bw];
-                  bgr->r = ( fore_r * p + bgr->r * (257 - p)) >> 8;
-                  bgr->g = ( fore_g * p + bgr->g * (257 - p)) >> 8;
-                  bgr->b = ( fore_b * p + bgr->b * (257 - p)) >> 8;
-                } while (++sy < bh);
-              }
-              me->push_image(bx, by, bw, bh, &p);
-            }
-            pixel += w;
-            if ((y += size_y) >= me->height()) break;
-          } while (--h);
-        }
-      }
-      me->endWrite();
-      return xAdvance;
-    }
-  };
-
-//----------------------------------------------------------------------------
-//----------------------------------------------------------------------------
-
-  template <class Base>
-  class LGFX_IMAGE_FORMAT_Support : public Base {
-  public:
-
-#if defined (ARDUINO)
- #if defined (FS_H)
-
-    inline void drawBmp(fs::FS &fs, const char *path, int32_t x=0, int32_t y=0) { drawBmpFile(fs, path, x, y); }
-    void drawBmpFile(fs::FS &fs, const char *path, int32_t x=0, int32_t y=0) {
-      FileWrapper file(fs);
-      drawBmpFile(&file, path, x, y);
-    }
-
-    inline bool drawJpgFile( fs::FS &fs, const char *path, int32_t x=0, int32_t y=0, int32_t maxWidth=0, int32_t maxHeight=0, int32_t offX=0, int32_t offY=0, jpeg_div_t scale=JPEG_DIV_NONE) {
-      FileWrapper file(fs);
-      return drawJpgFile(&file, path, x, y, maxWidth, maxHeight, offX, offY, scale);
-    }
-
-    inline bool drawPngFile( fs::FS &fs, const char *path, int32_t x = 0, int32_t y = 0, int32_t maxWidth = 0, int32_t maxHeight = 0, int32_t offX = 0, int32_t offY = 0, double scale = 1.0)
-    {
-      FileWrapper file(fs);
-      return drawPngFile( &file, path, x, y, maxWidth, maxHeight, offX, offY, scale);
-    }
-
- #endif
- #if defined (Stream_h)
-
-    inline bool drawJpg(Stream *dataSource, int32_t x=0, int32_t y=0, int32_t maxWidth=0, int32_t maxHeight=0, int32_t offX=0, int32_t offY=0, jpeg_div_t scale=JPEG_DIV_NONE) {
-      StreamWrapper data;
-      data.set(dataSource);
-      return draw_jpg(&data, x, y, maxWidth, maxHeight, offX, offY, scale);
-    }
-
-    inline void drawPng( Stream *dataSource, int32_t x = 0, int32_t y = 0, int32_t maxWidth = 0, int32_t maxHeight = 0, int32_t offX = 0, int32_t offY = 0, double scale = 1.0) {
-      StreamWrapper data;
-      data.set(dataSource);
-      return draw_png(&data, x, y, maxWidth, maxHeight, offX, offY, scale);
-    }
-
- #endif
-
-#elif defined (CONFIG_IDF_TARGET_ESP32)  // ESP-IDF
-
-    inline void drawBmpFile(const char *path, int32_t x, int32_t y) {
-      FileWrapper file;
-      drawBmpFile(&file, path, x, y);
-    }
-    inline bool drawJpgFile(const char *path, int32_t x=0, int32_t y=0, int32_t maxWidth=0, int32_t maxHeight=0, int32_t offX=0, int32_t offY=0, jpeg_div_t scale=JPEG_DIV_NONE) {
-      FileWrapper file;
-      return drawJpgFile(&file, path, x, y, maxWidth, maxHeight, offX, offY, scale);
-    }
-    inline bool drawPngFile( const char *path, int32_t x = 0, int32_t y = 0, int32_t maxWidth = 0, int32_t maxHeight = 0, int32_t offX = 0, int32_t offY = 0, double scale = 1.0)
-    {
-      FileWrapper file;
-      return drawPngFile( &file, path, x, y, maxWidth, maxHeight, offX, offY, scale);
-    }
-
-#endif
-
-    void drawBmp(const uint8_t *bmp_data, uint32_t bmp_len, int32_t x=0, int32_t y=0) {
-      PointerWrapper data;
-      data.set(bmp_data, bmp_len);
-      draw_bmp(&data, x, y);
-    }
-    bool drawJpg(const uint8_t *jpg_data, uint32_t jpg_len, int32_t x=0, int32_t y=0, int32_t maxWidth=0, int32_t maxHeight=0, int32_t offX=0, int32_t offY=0, jpeg_div_t scale=JPEG_DIV_NONE) {
-      PointerWrapper data;
-      data.set(jpg_data, jpg_len);
-      return draw_jpg(&data, x, y, maxWidth, maxHeight, offX, offY, scale);
-    }
-    bool drawPng( const uint8_t *png_data, uint32_t png_len, int32_t x = 0, int32_t y = 0, int32_t maxWidth = 0, int32_t maxHeight = 0, int32_t offX = 0, int32_t offY = 0, double scale = 1.0)
-    {
-      PointerWrapper data;
-      data.set(png_data, png_len);
-      return draw_png(&data, x, y, maxWidth, maxHeight, offX, offY, scale);
-    }
-
-  protected:
-
-    struct bitmap_header_t {
-      union {
-        uint8_t raw[54];
-        #pragma pack(1)
-        struct {
-          uint16_t bfType; 
-          uint32_t bfSize;
-          uint16_t bfReserved1;
-          uint16_t bfReserved2;
-          uint32_t bfOffBits;
-
-          uint32_t biSize; 
-          int32_t  biWidth;
-          int32_t  biHeight;
-          uint16_t biPlanes; 
-          uint16_t biBitCount;
-          uint32_t biCompression;
-          uint32_t biSizeImage; 
-          int32_t  biXPelsPerMeter;
-          int32_t  biYPelsPerMeter;
-          uint32_t biClrUsed; 
-          uint32_t biClrImportant;
-        };
-        #pragma pack()
-      };
-    };
-
-    bool load_bmp_header(DataWrapper* data, bitmap_header_t* result) {
-      data->read((uint8_t*)result, sizeof(bitmap_header_t));
-      return ((result->bfType == 0x4D42)   // bmp header "BM"
-           && (result->biPlanes == 1)  // bcPlanes always 1
-           && (result->biWidth > 0)
-           && (result->biHeight > 0)
-           && (result->biBitCount <= 32)
-           && (result->biBitCount != 0));
-    }
-
-    bool load_bmp_rle8(DataWrapper* data, uint8_t* linebuf, uint_fast16_t width) {
-      width = (width + 3) & ~3;
-      uint8_t code[2];
-      uint_fast16_t xidx = 0;
-      bool eol = false;
-      do {
-        data->read(code, 2);
-        if (code[0] == 0) {
-          switch (code[1]) {
-          case 0x00: // EOL
-          case 0x01: // EOB
-            eol = true;
-            break;
-
-          case 0x02: // move info  (not support)
-            return false;
-
-          default:
-            data->read(&linebuf[xidx], (code[1] + 1) & ~1); // word align
-            xidx += code[1];
-            break;
-          }
-        } else if (xidx + code[0] <= width) {
-          memset(&linebuf[xidx], code[1], code[0]);
-          xidx += code[0];
-        } else {
-          return false;
-        }
-      } while (!eol);
-      return true;
-    }
-
-    bool load_bmp_rle4(DataWrapper* data, uint8_t* linebuf, uint_fast16_t width) {
-      width = (width + 3) & ~3;
-      uint8_t code[2];
-      uint_fast16_t xidx = 0;
-      bool eol = false;
-      do {
-        data->read(code, 2);
-        if (code[0] == 0) {
-          switch (code[1]) {
-          case 0x00: // EOL
-          case 0x01: // EOB
-            eol = true;
-            break;
-
-          case 0x02: // move info  (not support)
-            return false;
-
-          default:  // 絶対モードデータ
-            {
-              int_fast16_t len = code[1];
-              int_fast16_t dbyte = ((int_fast16_t)code[1] + 1) >> 1;
-
-              data->read(&linebuf[(xidx + 1) >> 1], (dbyte + 1) & ~1); // word align
-              if (xidx & 1) {
-                linebuf[xidx >> 1] |= linebuf[(xidx >> 1) + 1] >> 4;
-                for (long i = 1; i < dbyte; ++i) {
-                  linebuf[((xidx + i) >> 1)] = (linebuf[((xidx + i) >> 1)    ] << 4)
-                                             |  linebuf[((xidx + i) >> 1) + 1] >> 4;
-                }
-              }
-              xidx += len;
-            }
-            break;
-          }
-        } else if (xidx + code[0] <= width) {
-          if (xidx & 1) {
-            linebuf[xidx >> 1] |= code[1] >> 4;
-            code[1] = (code[1] >> 4) | (code[1] << 4);
-          }
-          memset(&linebuf[(xidx + 1) >> 1], code[1], (code[0] + 1) >> 1);
-          xidx += code[0];
-          if (xidx & 1) linebuf[xidx >> 1] &= 0xF0;
-        } else {
-          return false;
-        }
-      } while (!eol);
-      return true;
-    }
-//*/
-  private:
-
-    void drawBmpFile(FileWrapper* file, const char *path, int32_t x=0, int32_t y=0) {
-      file->need_transaction &= this->isSPIShared();
-      if (file->need_transaction) this->endTransaction();
-      if (file->open(path, "rb")) {
-        draw_bmp(file, x, y);
-        file->close();
-      }
-      if (file->need_transaction && this->_transaction_count) { this->beginTransaction(); }
-    }
-
-    bool drawJpgFile(FileWrapper* file, const char *path, int32_t x, int32_t y, int32_t maxWidth, int32_t maxHeight, int32_t offX, int32_t offY, jpeg_div_t scale) {
-      bool res = false;
-      file->need_transaction &= this->isSPIShared();
-      if (file->need_transaction) this->endTransaction();
-      if (file->open(path, "rb")) {
-        res = draw_jpg(file, x, y, maxWidth, maxHeight, offX, offY, scale);
-        file->close();
-      }
-//    if (file->need_transaction && this->_transaction_count) { this->beginTransaction(); }
-      return res;
-    }
-
-    bool drawPngFile( FileWrapper* file, const char *path, int32_t x, int32_t y, int32_t maxWidth, int32_t maxHeight, int32_t offX, int32_t offY, double scale)
-    {
-      bool res = false;
-      file->need_transaction &= this->isSPIShared();
-      if (file->need_transaction) this->endTransaction();
-      if (file->open(path, "rb")) {
-        res = draw_png(file, x, y, maxWidth, maxHeight, offX, offY, scale);
-        file->close();
-      }
-      if (file->need_transaction && this->_transaction_count) { this->beginTransaction(); }
-      return res;
-    }
-
-    void draw_bmp(DataWrapper* data, int32_t x, int32_t y) {
-      if ((x >= this->_width) || (y >= this->_height)) return;
-
-      bitmap_header_t bmpdata;
-      if (!load_bmp_header(data, &bmpdata)
-       || (bmpdata.biCompression > 3)) {
-        return;
-      }
-
-      //uint32_t startTime = millis();
-      uint32_t seekOffset = bmpdata.bfOffBits;
-      int32_t w = bmpdata.biWidth;
-      int32_t h = bmpdata.biHeight;  // bcHeight Image height (pixels)
-      uint_fast16_t bpp = bmpdata.biBitCount; // 24 bcBitCount 24=RGB24bit
-
-        //If the value of Height is positive, the image data is from bottom to top
-        //If the value of Height is negative, the image data is from top to bottom.
-      int32_t flow = (h < 0) ? 1 : -1;
-      if (h < 0) h = -h;
-      else y += h - 1;
-
-      argb8888_t *palette = nullptr;
-      if (bpp <= 8) {
-        palette = new argb8888_t[1 << bpp];
-        data->seek(bmpdata.biSize + 14);
-        data->read((uint8_t*)palette, (1 << bpp)*sizeof(argb8888_t)); // load palette
-      }
-
-      data->seek(seekOffset);
-
-      auto dst_depth = this->_write_conv.depth;
-      uint32_t buffersize = ((w * bpp + 31) >> 5) << 2;  // readline 4Byte align.
-      uint8_t lineBuffer[buffersize + 4];
-      pixelcopy_t p(lineBuffer, dst_depth, (color_depth_t)bpp, this->_palette_count, palette);
-      p.no_convert = false;
-      if (8 >= bpp && !this->_palette_count) {
-        p.fp_copy = pixelcopy_t::get_fp_palettecopy<argb8888_t>(dst_depth);
-      } else {
-        if (bpp == 16) {
-          p.fp_copy = pixelcopy_t::get_fp_normalcopy<rgb565_t>(dst_depth);
-        } else if (bpp == 24) {
-          p.fp_copy = pixelcopy_t::get_fp_normalcopy<rgb888_t>(dst_depth);
-        } else if (bpp == 32) {
-          p.fp_copy = pixelcopy_t::get_fp_normalcopy<argb8888_t>(dst_depth);
-        }
-      }
-
-      auto nt = data->need_transaction;
-      if (bmpdata.biCompression == 1) {
-        do {
-          if (nt) this->endTransaction();
-          load_bmp_rle8(data, lineBuffer, w);
-          if (nt) this->beginTransaction();
-          this->push_image(x, y, w, 1, &p);
-          y += flow;
-        } while (--h);
-      } else
-      if (bmpdata.biCompression == 2) {
-        do {
-          if (nt) this->endTransaction();
-          load_bmp_rle4(data, lineBuffer, w);
-          if (nt) this->beginTransaction();
-          this->push_image(x, y, w, 1, &p);
-          y += flow;
-        } while (--h);
-      } else {
-        do {
-          if (nt) this->endTransaction();
-          data->read(lineBuffer, buffersize);
-          if (nt) this->beginTransaction();
-          this->push_image(x, y, w, 1, &p);
-          y += flow;
-        } while (--h);
-      }
-      if (palette) delete[] palette;
-      //Serial.print("Loaded in "); Serial.print(millis() - startTime);   Serial.println(" ms");
-    }
-
-
-protected:
-
-#if !defined (__LGFX_TJPGDEC_H__)
-
-    bool draw_jpg(DataWrapper* data, int32_t x, int32_t y, int32_t maxWidth, int32_t maxHeight, int32_t offX, int32_t offY, jpeg_div_t scale)
-    {
-//ESP_LOGI("LGFX","drawJpg need include utility/tjpgd.h");
-      return false;
-    }
-
-#else
-
-    bool draw_jpg(DataWrapper* data, int32_t x, int32_t y, int32_t maxWidth, int32_t maxHeight, int32_t offX, int32_t offY, jpeg_div_t scale)
-    {
-      draw_jpg_info_t jpeg;
-      pixelcopy_t pc(nullptr, this->getColorDepth(), bgr888_t::depth, this->hasPalette());
-      jpeg.pc = &pc;
-      jpeg.lgfx = this;
-      jpeg.data = data;
-      jpeg.x = x - offX;
-      jpeg.y = y - offY;
-
-      //TJpgD jpegdec;
-      lgfxJdec jpegdec;
-
-      static constexpr uint16_t sz_pool = 3100;
-      uint8_t *pool = (uint8_t*)heap_alloc_dma(sz_pool);
-      if (!pool) {
-//        ESP_LOGE("LGFX","memory allocation failure");
-        return false;
-      }
-
-      auto jres = lgfx_jd_prepare(&jpegdec, jpg_read_data, pool, sz_pool, &jpeg);
-
-      if (jres != JDR_OK) {
-//ESP_LOGE("LGFX","jpeg prepare error:%x", jres);
-        heap_free(pool);
-        return false;
-      }
-
-      if (!maxWidth) maxWidth = this->width();
-      auto cl = this->_clip_l;
-      if (0 > x - cl) { maxWidth += x - cl; x = cl; }
-      auto cr = this->_clip_r + 1;
-      if (maxWidth > (cr - x)) maxWidth = (cr - x);
-
-      if (!maxHeight) maxHeight = this->height();
-      auto ct = this->_clip_t;
-      if (0 > y - ct) { maxHeight += y - ct; y = ct; }
-      auto cb = this->_clip_b + 1;
-      if (maxHeight > (cb - y)) maxHeight = (cb - y);
-
-      if (maxWidth > 0 && maxHeight > 0) {
-        this->setClipRect(x, y, maxWidth, maxHeight);
-        this->startWrite();
-        jres = lgfx_jd_decomp(&jpegdec, jpg_push_image, scale);
-
-        this->_clip_l = cl;
-        this->_clip_t = ct;
-        this->_clip_r = cr-1;
-        this->_clip_b = cb-1;
-        this->endWrite();
-      }
-      heap_free(pool);
-
-      if (jres != JDR_OK) {
-//ESP_LOGE("LGFX","jpeg decomp error:%x", jres);
-        return false;
-      }
-      return true;
-    }
-
-    struct draw_jpg_info_t {
-      int32_t x;
-      int32_t y;
-      DataWrapper *data;
-      LGFXBase *lgfx;
-      pixelcopy_t *pc;
-    };
-
-    static uint32_t jpg_read_data(lgfxJdec  *decoder, uint8_t *buf, uint32_t len) {
-      auto jpeg = (draw_jpg_info_t *)decoder->device;
-      auto data = (DataWrapper*)jpeg->data;
-      auto res = len;
-      if (data->need_transaction) jpeg->lgfx->endTransaction();
-      if (buf) {
-        res = data->read(buf, len);
-      } else {
-        data->skip(len);
-      }
-      return res;
-    }
-
-    static uint32_t jpg_push_image(lgfxJdec *decoder, void *bitmap, JRECT *rect) {
-      draw_jpg_info_t *jpeg = (draw_jpg_info_t *)decoder->device;
-      jpeg->pc->src_data = bitmap;
-      auto data = (DataWrapper*)jpeg->data;
-      if (data->need_transaction) jpeg->lgfx->beginTransaction();
-      jpeg->lgfx->push_image( jpeg->x + rect->left
-                           , jpeg->y + rect->top 
-                           , rect->right  - rect->left + 1
-                           , rect->bottom - rect->top + 1
-                           , jpeg->pc
-                           , false);
-      return 1;
-    }
-
-#endif
-
-
-#ifndef __LGFX_PNGLE_H__
-
-    bool draw_png(DataWrapper* data, int32_t x, int32_t y, int32_t maxWidth, int32_t maxHeight, int32_t offX, int32_t offY, double scale)
-    {
-//ESP_LOGI("LGFX","drawPng need include utility/pngle.h");
-      return false;
-    }
-
-#else
-
-    struct png_file_decoder_t {
-      int32_t x;
-      int32_t y;
-      int32_t offX;
-      int32_t offY;
-      int32_t maxWidth;
-      int32_t maxHeight;
-      double scale;
-      bgr888_t* lineBuffer;
-      pixelcopy_t *pc;
-      LGFXBase *lgfx;
-      int32_t last_y;
-      int32_t scale_y0;
-      int32_t scale_y1;
-    };
-
-    static bool png_ypos_update(png_file_decoder_t *p, uint32_t y)
-    {
-      p->scale_y0 = ceil( y      * p->scale) - p->offY;
-      if (p->scale_y0 < 0) p->scale_y0 = 0;
-      p->scale_y1 = ceil((y + 1) * p->scale) - p->offY;
-      if (p->scale_y1 > p->maxHeight) p->scale_y1 = p->maxHeight;
-      return (p->scale_y0 < p->scale_y1);
-    }
-
-    static void png_post_line(png_file_decoder_t *p, uint32_t y)
-    {
-      int32_t h = p->scale_y1 - p->scale_y0;
-      if (0 < h)
-        p->lgfx->push_image(p->x, p->y + p->scale_y0, p->maxWidth, h, p->pc, true);
-    }
-
-    static void png_prepare_line(png_file_decoder_t *p, uint32_t y)
-    {
-      p->last_y = y;
-      if (png_ypos_update(p, y))      // read next line
-        p->lgfx->readRectRGB(p->x, p->y + p->scale_y0, p->maxWidth, p->scale_y1 - p->scale_y0, p->lineBuffer);
-    }
-
-    static void png_done_callback(pngle_t *pngle)
-    {
-      auto p = (png_file_decoder_t *)lgfx_pngle_get_user_data(pngle);
-      png_post_line(p, p->last_y);
-    }
-
-    static void png_draw_normal_callback(pngle_t *pngle, uint32_t x, uint32_t y, uint8_t rgba[4])
-    {
-      auto p = (png_file_decoder_t*)lgfx_pngle_get_user_data(pngle);
-
-      int32_t t = y - p->offY;
-      if (t < 0 || t >= p->maxHeight) return;
-
-      int32_t l = x - p->offX;
-      if (l < 0 || l >= p->maxWidth) return;
-
-      p->lgfx->setColor(color888(rgba[0], rgba[1], rgba[2]));
-      p->lgfx->writeFillRectPreclipped(p->x + l, p->y + t, 1, 1);
-    }
-
-    static void png_draw_normal_scale_callback(pngle_t *pngle, uint32_t x, uint32_t y, uint8_t rgba[4])
-    {
-      auto p = (png_file_decoder_t*)lgfx_pngle_get_user_data(pngle);
-
-      if (y != p->last_y) {
-        p->last_y = y;
-        png_ypos_update(p, y);
-      }
-
-      int32_t t = p->scale_y0;
-      int32_t h = p->scale_y1 - t;
-      if (h <= 0) return;
-
-      int32_t l = ceil( x      * p->scale) - p->offX;
-      if (l < 0) l = 0;
-      int32_t r = ceil((x + 1) * p->scale) - p->offX;
-      if (r > p->maxWidth) r = p->maxWidth;
-      if (l >= r) return;
-
-      p->lgfx->setColor(color888(rgba[0], rgba[1], rgba[2]));
-      p->lgfx->writeFillRectPreclipped(p->x + l, p->y + t, r - l, h);
-    }
-
-    static void png_draw_alpha_callback(pngle_t *pngle, uint32_t x, uint32_t y, uint8_t rgba[4])
-    {
-      auto p = (png_file_decoder_t*)lgfx_pngle_get_user_data(pngle);
-      if (y != p->last_y) {
-        png_post_line(p, p->last_y);
-        png_prepare_line(p, y);
-      }
-
-      if (p->scale_y0 >= p->scale_y1) return;
-
-      int32_t l = ( x      ) - p->offX;
-      if (l < 0) l = 0;
-      int32_t r = ((x + 1) ) - p->offX;
-      if (r > p->maxWidth) r = p->maxWidth;
-      if (l >= r) return;
-
-      if (rgba[3] == 255) {
-        memcpy(&p->lineBuffer[l], rgba, 3);
-      } else {
-        auto data = &p->lineBuffer[l];
-        uint_fast8_t alpha = rgba[3] + 1;
-        data->r = (rgba[0] * alpha + data->r * (257 - alpha)) >> 8;
-        data->g = (rgba[1] * alpha + data->g * (257 - alpha)) >> 8;
-        data->b = (rgba[2] * alpha + data->b * (257 - alpha)) >> 8;
-      }
-    }
-
-    static void png_draw_alpha_scale_callback(pngle_t *pngle, uint32_t x, uint32_t y, uint8_t rgba[4])
-    {
-      auto p = (png_file_decoder_t*)lgfx_pngle_get_user_data(pngle);
-      if (y != p->last_y) {
-        png_post_line(p, p->last_y);
-        png_prepare_line(p, y);
-      }
-
-      int32_t b = p->scale_y1 - p->scale_y0;
-      if (b <= 0) return;
-
-      int32_t l = ceil( x      * p->scale) - p->offX;
-      if (l < 0) l = 0;
-      int32_t r = ceil((x + 1) * p->scale) - p->offX;
-      if (r > p->maxWidth) r = p->maxWidth;
-      if (l >= r) return;
-
-      if (rgba[3] == 255) {
-        int32_t i = l;
-        do {
-          for (int32_t j = 0; j < b; ++j) {
-            auto data = &p->lineBuffer[i + j * p->maxWidth];
-            memcpy(data, rgba, 3);
-          }
-        } while (++i < r);
-      } else {
-        uint_fast8_t alpha = rgba[3] + 1;
-        int32_t i = l;
-        do {
-          for (int32_t j = 0; j < b; ++j) {
-            auto data = &p->lineBuffer[i + j * p->maxWidth];
-            data->r = (rgba[0] * alpha + data->r * (257 - alpha)) >> 8;
-            data->g = (rgba[1] * alpha + data->g * (257 - alpha)) >> 8;
-            data->b = (rgba[2] * alpha + data->b * (257 - alpha)) >> 8;
-          }
-        } while (++i < r);
-      }
-    }
-
-    static void png_init_callback(pngle_t *pngle, uint32_t w, uint32_t h, uint_fast8_t hasTransparent)
-    {
-//    auto ihdr = lgfx_pngle_get_ihdr(pngle);
-
-      auto p = (png_file_decoder_t*)lgfx_pngle_get_user_data(pngle);
-
-      if (p->scale != 1.0) {
-        w = ceil(w * p->scale);
-        h = ceil(h * p->scale);
-      }
-
-      int32_t ww = w - abs(p->offX);
-      if (p->maxWidth > ww) p->maxWidth = ww;
-      if (p->maxWidth < 0) return;
-      if (p->offX < 0) { p->offX = 0; }
-
-      int32_t hh = h - abs(p->offY);
-      if (p->maxHeight > hh) p->maxHeight = hh;
-      if (p->maxHeight < 0) return;
-      if (p->offY < 0) { p->offY = 0; }
-
-      if (hasTransparent) { // need pixel read ?
-        p->lineBuffer = (bgr888_t*)heap_alloc_dma(sizeof(bgr888_t) * p->maxWidth * ceil(p->scale));
-        p->pc->src_data = p->lineBuffer;
-        png_prepare_line(p, 0);
-        lgfx_pngle_set_done_callback(pngle, png_done_callback);
-
-        if (p->scale == 1.0) {
-          lgfx_pngle_set_draw_callback(pngle, png_draw_alpha_callback);
-        } else {
-          lgfx_pngle_set_draw_callback(pngle, png_draw_alpha_scale_callback);
-        }
-      } else {
-        if (p->scale == 1.0) {
-          lgfx_pngle_set_draw_callback(pngle, png_draw_normal_callback);
-        } else {
-          p->last_y = 0;
-          png_ypos_update(p, 0);
-          lgfx_pngle_set_draw_callback(pngle, png_draw_normal_scale_callback);
-        }
-        return;
-      }
-    }
-
-    bool draw_png(DataWrapper* data, int32_t x, int32_t y, int32_t maxWidth, int32_t maxHeight, int32_t offX, int32_t offY, double scale)
-    {
-      if (!maxHeight) maxHeight = INT32_MAX;
-      auto ct = this->_clip_t;
-      if (0 > y - ct) { maxHeight += y - ct; offY -= y - ct; y = ct; }
-      if (0 > offY) { y -= offY; maxHeight += offY; offY = 0; }
-      auto cb = this->_clip_b + 1;
-      if (maxHeight > (cb - y)) maxHeight = (cb - y);
-      if (maxHeight < 0) return true;
-
-      if (!maxWidth) maxWidth = INT32_MAX;
-      auto cl = this->_clip_l;
-      if (0 > x - cl) { maxWidth += x - cl; offX -= x - cl; x = cl; }
-      if (0 > offX) { x -= offX; maxWidth  += offX; offX = 0; }
-      auto cr = this->_clip_r + 1;
-      if (maxWidth > (cr - x)) maxWidth = (cr - x);
-      if (maxWidth < 0) return true;
-
-      png_file_decoder_t png;
-      png.x = x;
-      png.y = y;
-      png.offX = offX;
-      png.offY = offY;
-      png.maxWidth = maxWidth;
-      png.maxHeight = maxHeight;
-      png.scale = scale;
-      png.lgfx = this;
-      png.lineBuffer = nullptr;
-
-      pixelcopy_t pc(nullptr, this->getColorDepth(), bgr888_t::depth, this->_palette_count);
-      png.pc = &pc;
-
-      pngle_t *pngle = lgfx_pngle_new();
-
-      lgfx_pngle_set_user_data(pngle, &png);
-
-      lgfx_pngle_set_init_callback(pngle, png_init_callback);
-
-      // Feed data to pngle
-      uint8_t buf[512];
-      int remain = 0;
-      int len;
-      bool res = true;
-
-      len = data->read(buf, sizeof(buf));
-
-      if (data->need_transaction && this->_transaction_count) this->beginTransaction();
-      this->startWrite();
-      while (len > 0) {
-
-        int fed = lgfx_pngle_feed(pngle, buf, remain + len);
-
-        if (fed < 0) {
-//ESP_LOGE("LGFX", "[pngle error] %s", lgfx_pngle_error(pngle));
-          res = false;
-          break;
-        }
-
-        remain = remain + len - fed;
-        if (remain > 0) memmove(buf, buf + fed, remain);
-        if (data->need_transaction) this->endTransaction();
-        len = data->read(buf + remain, sizeof(buf) - remain);
-        if (data->need_transaction) this->beginTransaction();
-      }
-      this->endWrite();
-      if (png.lineBuffer) {
-        this->waitDMA();
-        heap_free(png.lineBuffer);
-      }
-      lgfx_pngle_destroy(pngle);
-      return res;
-    }
-
-#endif
-
-
-  };
-//----------------------------------------------------------------------------
-
-//----------------------------------------------------------------------------
 }
-
-class LovyanGFX : public
-   lgfx::LGFX_VLWFont_Support<
-    lgfx::LGFX_IMAGE_FORMAT_Support<
-     lgfx::LGFXBase
-    >
-   >
-{};
-//----------------------------------------------------------------------------
-//----------------------------------------------------------------------------
 
 #endif
