+++ src/lgfx/lgfx_common.hpp
@@ -36,27 +36,6 @@ namespace lgfx
 {
   static constexpr float deg_to_rad = 0.017453292519943295769236907684886;
 
-  enum textdatum_t : uint8_t
-  //  0:left   1:centre   2:right
-  //  0:top    4:middle   8:bottom   16:baseline
-  { top_left        =  0  // Top left (default)
-  , top_center      =  1  // Top center
-  , top_centre      =  1  // Top center
-  , top_right       =  2  // Top right
-  , middle_left     =  4  // Middle left
-  , middle_center   =  5  // Middle center
-  , middle_centre   =  5  // Middle center
-  , middle_right    =  6  // Middle right
-  , bottom_left     =  8  // Bottom left
-  , bottom_center   =  9  // Bottom center
-  , bottom_centre   =  9  // Bottom center
-  , bottom_right    = 10  // Bottom right
-  , baseline_left   = 16  // Baseline left (Line the 'A' character would sit on)
-  , baseline_center = 17  // Baseline center
-  , baseline_centre = 17  // Baseline center
-  , baseline_right  = 18  // Baseline right
-  };
-
   enum color_depth_t : uint8_t
   { palette_1bit   =  1 //   2 color
   , palette_2bit   =  2 //   4 color
